Add RuntimeState, fix '>>> ' issue, enable most directives, improve failure repr, reorganize modules, improve __main__
