+++ xdoctest/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 # flake8: noqa
 __version__ = '0.0.12'  # nocover
+from xdoctest.exceptions import ExitTestException
 from xdoctest.runner import doctest_module  # nocover
