+++ xdoctest/constants.py
@@ -1,0 +1,1 @@
+NOT_EVALED = object()  # sentinal value
